Fix broken wsgify.middleware decorator functionality.
wsgify.__call__ was not passing correct arguments to the middleware function.

It appears that it worked before because there are doctests in test_dec.txt that demo this feature.
Also updated docstring to match actual argument order in middleware functions: req, app, ...

PS: wsgify.__call__'s kw argument doesn't seem to be extremely useful, perhaps it is a part of the debris left over by a refactoring or two.